refactor: apply pr-review to remove transient properties from status policy
